Remove "src" folder; the "HTML Engine" and "Java Engine" now have their own repos.
  http://github.com/TooTallNate/SGF-HTML-Engine
  http://github.com/TooTallNate/SGF-Java-Engine

This main repo will shift to being the home of the "specification" itself. It will contain the raw code to generate the 'html', 'man', and maybe at some point 'pdf' versions of the "specification".

It will remain to hold the homepage for `http://www.simplegameframework.com`, and hold 'example' games that should be compliant by any implementation of the Simple Game Framework.